add cashier to header
